split up user db and data
