+++ src/mocks/handlers/user/item.handlers.ts
@@ -1,5 +1,5 @@
 const { rest } = window.MockServiceWorker;
-import { umbUsersData } from '../../data/user.data.js';
+import { umbUsersData } from '../../data/user/user.db.js';
 import { UMB_SLUG } from './slug.js';
 import { umbracoPath } from '@umbraco-cms/backoffice/utils';
